+++ src/config/config.ts
@@ -1,7 +1,8 @@
 export const WEB3_MAIL_DAPP_ADDRESS = 'web3mail.apps.iexec.eth';
-export const WORKERPOOL_ADDRESS = 'prod-v8-bellecour.main.pools.iexec.eth';
+export const PROD_WORKERPOOL_ADDRESS = 'prod-v8-bellecour.main.pools.iexec.eth';
 export const DATAPROTECTOR_SUBGRAPH_ENDPOINT =
   'https://thegraph-product.iex.ec/subgraphs/name/bellecour/dataprotector';
+export const MAX_DESIRED_DATA_ORDER_PRICE = 0;
 export const MAX_DESIRED_APP_ORDER_PRICE = 0;
 export const MAX_DESIRED_WORKERPOOL_ORDER_PRICE = 0;
 export const DEFAULT_CONTENT_TYPE = 'text/plain';
@@ -9,3 +10,4 @@ export const IPFS_UPLOAD_URL = '/dns4/ipfs-upload.v8-bellecour.iex.ec/https';
 export const DEFAULT_IPFS_GATEWAY = 'https://ipfs-gateway.v8-bellecour.iex.ec';
 export const WHITELIST_SMART_CONTRACT_ADDRESS =
   '0x781482C39CcE25546583EaC4957Fb7Bf04C277D2';
+export const ANY_DATASET_ADDRESS = 'any';