+++ src/utils/validators.ts
@@ -1,15 +1,14 @@
-import { utils } from 'ethers';
-import { ValidationError, string } from 'yup';
-
-const { isAddress } = utils;
+import { isAddress } from 'ethers';
+import { ValidationError, number, string } from 'yup';
 
 export const throwIfMissing = (): never => {
   throw new ValidationError('Missing parameter');
 };
 
-const isUndefined = (value: any) => value === undefined;
+const isUndefined = (value: unknown) => value === undefined;
 const isAddressTest = (value: string) => isAddress(value);
-const isEnsTest = (value: string) => value.endsWith('.eth') && value.length > 6;
+export const isEnsTest = (value: string) =>
+  value.endsWith('.eth') && value.length > 6;
 
 export const addressOrEnsSchema = () =>
   string()
@@ -18,6 +17,15 @@ export const addressOrEnsSchema = () =>
       'is-address-or-ens',
       '${path} should be an ethereum address or a ENS name',
       (value) => isUndefined(value) || isAddressTest(value) || isEnsTest(value)
+    );
+
+export const addressSchema = () =>
+  string()
+    .transform((value: string) => value?.toLowerCase() || value)
+    .test(
+      'is-address',
+      '${path} should be an ethereum address',
+      (value) => isUndefined(value) || isAddressTest(value)
     );
 
 // 78 char length for email subject (rfc2822)
@@ -34,3 +42,9 @@ export const contentTypeSchema = () =>
 
 // Minimum of 3 characters and max of 20 to avoid sender being flagged as spam
 export const senderNameSchema = () => string().trim().min(3).max(20).optional();
+
+// Used to identify the email campaign, minimum of 3 characters and max of 10
+export const labelSchema = () => string().trim().min(3).max(10).optional();
+
+export const positiveNumberSchema = () =>
+  number().integer().min(0).typeError('${path} must be a non-negative number');