+++ src/web3mail/IExecWeb3mail.ts
@@ -1,49 +1,91 @@
-import { providers } from 'ethers';
+import { Eip1193Provider } from 'ethers';
 import { IExec } from 'iexec';
-import { IExecConfigOptions } from 'iexec/IExecConfig';
+import { fetchUserContacts } from './fetchUserContacts.js';
 import { fetchMyContacts } from './fetchMyContacts.js';
 import { sendEmail } from './sendEmail.js';
 import {
   Contact,
-  FetchContactsParams,
+  FetchUserContactsParams,
   SendEmailParams,
-  SendEmailResponse,
   Web3SignerProvider,
+  AddressOrENS,
+  Web3MailConfigOptions,
 } from './types.js';
 import { GraphQLClient } from 'graphql-request';
-import { DATAPROTECTOR_SUBGRAPH_ENDPOINT } from '../config/config.js';
+import {
+  WEB3_MAIL_DAPP_ADDRESS,
+  IPFS_UPLOAD_URL,
+  DEFAULT_IPFS_GATEWAY,
+  DATAPROTECTOR_SUBGRAPH_ENDPOINT,
+  WHITELIST_SMART_CONTRACT_ADDRESS,
+} from '../config/config.js';
 
 export class IExecWeb3mail {
-  fetchMyContacts: (args?: FetchContactsParams) => Promise<Contact[]>;
-  sendEmail: (args: SendEmailParams) => Promise<SendEmailResponse>;
+  private iexec: IExec;
+
+  private ipfsNode: string;
+
+  private ipfsGateway: string;
+
+  private dataProtectorSubgraph: string;
+
+  private dappAddressOrENS: AddressOrENS;
+
+  private dappWhitelistAddress: AddressOrENS;
+
+  private graphQLClient: GraphQLClient;
 
   constructor(
-    ethProvider: providers.ExternalProvider | Web3SignerProvider,
-    options?: {
-      iexecOptions?: IExecConfigOptions;
-    }
+    ethProvider: Eip1193Provider | Web3SignerProvider,
+    options?: Web3MailConfigOptions
   ) {
-    let iexec: IExec;
-    let graphQLClient: GraphQLClient;
     try {
-      iexec = new IExec({ ethProvider }, options?.iexecOptions);
+      this.iexec = new IExec({ ethProvider }, options?.iexecOptions);
     } catch (e) {
       throw Error('Unsupported ethProvider');
     }
 
     try {
-      graphQLClient = new GraphQLClient(DATAPROTECTOR_SUBGRAPH_ENDPOINT);
+      this.dataProtectorSubgraph =
+        options?.dataProtectorSubgraph || DATAPROTECTOR_SUBGRAPH_ENDPOINT;
+      this.graphQLClient = new GraphQLClient(this.dataProtectorSubgraph);
     } catch (e) {
       throw Error('Impossible to create GraphQLClient');
     }
 
-    this.fetchMyContacts = (args?: FetchContactsParams) =>
-      fetchMyContacts({ ...args, iexec, graphQLClient });
-    this.sendEmail = (args: SendEmailParams) =>
-      sendEmail({
-        ...args,
-        iexec,
-        graphQLClient,
-      });
+    this.dappAddressOrENS = options?.dappAddressOrENS || WEB3_MAIL_DAPP_ADDRESS;
+    this.ipfsNode = options?.ipfsNode || IPFS_UPLOAD_URL;
+    this.ipfsGateway = options?.ipfsGateway || DEFAULT_IPFS_GATEWAY;
+    this.dappWhitelistAddress =
+      options?.dappWhitelistAddress || WHITELIST_SMART_CONTRACT_ADDRESS;
   }
+
+  fetchMyContacts = () =>
+    fetchMyContacts({
+      iexec: this.iexec,
+      graphQLClient: this.graphQLClient,
+      dappAddressOrENS: this.dappAddressOrENS,
+      dappWhitelistAddress: this.dappWhitelistAddress,
+    });
+
+  fetchUserContacts(args?: FetchUserContactsParams): Promise<Contact[]> {
+    return fetchUserContacts({
+      ...args,
+      iexec: this.iexec,
+      graphQLClient: this.graphQLClient,
+      dappAddressOrENS: this.dappAddressOrENS,
+      dappWhitelistAddress: this.dappWhitelistAddress,
+    });
+  }
+
+  sendEmail = (args: SendEmailParams) =>
+    sendEmail({
+      ...args,
+      iexec: this.iexec,
+      ipfsNode: this.ipfsNode,
+      ipfsGateway: this.ipfsGateway,
+      dappAddressOrENS: this.dappAddressOrENS,
+      dappWhitelistAddress: this.dappWhitelistAddress,
+      graphQLClient: this.graphQLClient,
+    });
 }