+++ src/web3mail/sendEmail.ts
@@ -1,39 +1,59 @@
 import {
   DEFAULT_CONTENT_TYPE,
   MAX_DESIRED_APP_ORDER_PRICE,
+  MAX_DESIRED_DATA_ORDER_PRICE,
   MAX_DESIRED_WORKERPOOL_ORDER_PRICE,
-  WEB3_MAIL_DAPP_ADDRESS,
-  WORKERPOOL_ADDRESS,
+  PROD_WORKERPOOL_ADDRESS,
 } from '../config/config.js';
 import { WorkflowError } from '../utils/errors.js';
 import { generateSecureUniqueId } from '../utils/generateUniqueId.js';
+import * as ipfs from '../utils/ipfs-service.js';
 import { checkProtectedDataValidity } from '../utils/subgraphQuery.js';
 import {
   addressOrEnsSchema,
+  addressSchema,
   contentTypeSchema,
   emailContentSchema,
   emailSubjectSchema,
+  labelSchema,
+  positiveNumberSchema,
   senderNameSchema,
   throwIfMissing,
 } from '../utils/validators.js';
 import {
+  DappAddressConsumer,
+  DappWhitelistAddressConsumer,
   IExecConsumer,
+  IpfsGatewayConfigConsumer,
+  IpfsNodeConfigConsumer,
   SendEmailParams,
   SendEmailResponse,
   SubgraphConsumer,
 } from './types.js';
-import * as ipfs from './../utils/ipfs-service.js';
 
 export const sendEmail = async ({
   graphQLClient = throwIfMissing(),
   iexec = throwIfMissing(),
+  workerpoolAddressOrEns = PROD_WORKERPOOL_ADDRESS,
+  dappAddressOrENS,
+  dappWhitelistAddress,
+  ipfsNode,
+  ipfsGateway,
   emailSubject,
   emailContent,
   contentType = DEFAULT_CONTENT_TYPE,
+  label,
+  dataMaxPrice = MAX_DESIRED_DATA_ORDER_PRICE,
+  appMaxPrice = MAX_DESIRED_APP_ORDER_PRICE,
+  workerpoolMaxPrice = MAX_DESIRED_WORKERPOOL_ORDER_PRICE,
   senderName,
   protectedData,
 }: IExecConsumer &
   SubgraphConsumer &
+  DappAddressConsumer &
+  DappWhitelistAddressConsumer &
+  IpfsNodeConfigConsumer &
+  IpfsGatewayConfigConsumer &
   SendEmailParams): Promise<SendEmailResponse> => {
   try {
     const vDatasetAddress = addressOrEnsSchema()
@@ -55,6 +75,28 @@ export const sendEmail = async ({
     const vSenderName = senderNameSchema()
       .label('senderName')
       .validateSync(senderName);
+    const vLabel = labelSchema().label('label').validateSync(label);
+    const vWorkerpoolAddressOrEns = addressOrEnsSchema()
+      .required()
+      .label('WorkerpoolAddressOrEns')
+      .validateSync(workerpoolAddressOrEns);
+    const vDappAddressOrENS = addressOrEnsSchema()
+      .required()
+      .label('dappAddressOrENS')
+      .validateSync(dappAddressOrENS);
+    const vDappWhitelistAddress = addressSchema()
+      .required()
+      .label('dappWhitelistAddress')
+      .validateSync(dappWhitelistAddress);
+    const vDataMaxPrice = positiveNumberSchema()
+      .label('dataMaxPrice')
+      .validateSync(dataMaxPrice);
+    const vAppMaxPrice = positiveNumberSchema()
+      .label('appMaxPrice')
+      .validateSync(appMaxPrice);
+    const vWorkerpoolMaxPrice = positiveNumberSchema()
+      .label('workerpoolMaxPrice')
+      .validateSync(workerpoolMaxPrice);
 
     // Check protected data validity through subgraph
     const isValidProtectedData = await checkProtectedDataValidity(
@@ -75,67 +117,83 @@ export const sendEmail = async ({
       await iexec.storage.pushStorageToken(token);
     }
 
-    // Fetch dataset order
-    const datasetOrderbook = await iexec.orderbook.fetchDatasetOrderbook(
-      vDatasetAddress,
-      {
-        app: WEB3_MAIL_DAPP_ADDRESS,
-        requester: requesterAddress,
-      }
-    );
-    const datasetorder = datasetOrderbook?.orders[0]?.order;
+    const [
+      datasetorderForApp,
+      datasetorderForWhitelist,
+      apporder,
+      workerpoolorder,
+    ] = await Promise.all([
+      // Fetch dataset order for web3mail app
+      iexec.orderbook
+        .fetchDatasetOrderbook(vDatasetAddress, {
+          app: dappAddressOrENS,
+          requester: requesterAddress,
+        })
+        .then((datasetOrderbook) => {
+          const desiredPriceDataOrderbook = datasetOrderbook.orders.filter(
+            (order) => order.order.datasetprice <= vDataMaxPrice
+          );
+          return desiredPriceDataOrderbook[0]?.order; // may be undefined
+        }),
+      // Fetch dataset order for web3mail whitelist
+      iexec.orderbook
+        .fetchDatasetOrderbook(vDatasetAddress, {
+          app: vDappWhitelistAddress,
+          requester: requesterAddress,
+        })
+        .then((datasetOrderbook) => {
+          const desiredPriceDataOrderbook = datasetOrderbook.orders.filter(
+            (order) => order.order.datasetprice <= vDataMaxPrice
+          );
+          return desiredPriceDataOrderbook[0]?.order; // may be undefined
+        }),
+      // Fetch app order
+      iexec.orderbook
+        .fetchAppOrderbook(dappAddressOrENS, {
+          minTag: ['tee', 'scone'],
+          maxTag: ['tee', 'scone'],
+          workerpool: workerpoolAddressOrEns,
+        })
+        .then((appOrderbook) => {
+          const desiredPriceAppOrderbook = appOrderbook.orders.filter(
+            (order) => order.order.appprice <= vAppMaxPrice
+          );
+          const desiredPriceAppOrder = desiredPriceAppOrderbook[0]?.order;
+          if (!desiredPriceAppOrder) {
+            throw new Error('No App order found for the desired price');
+          }
+          return desiredPriceAppOrder;
+        }),
+      // Fetch workerpool order
+      iexec.orderbook
+        .fetchWorkerpoolOrderbook({
+          workerpool: workerpoolAddressOrEns,
+          app: dappAddressOrENS,
+          dataset: vDatasetAddress,
+          minTag: ['tee', 'scone'],
+          maxTag: ['tee', 'scone'],
+          category: 0,
+        })
+        .then((workerpoolOrderbook) => {
+          const desiredPriceWorkerpoolOrderbook =
+            workerpoolOrderbook.orders.filter(
+              (order) => order.order.workerpoolprice <= vWorkerpoolMaxPrice
+            );
+          const randomIndex = Math.floor(
+            Math.random() * desiredPriceWorkerpoolOrderbook.length
+          );
+          const desiredPriceWorkerpoolOrder =
+            desiredPriceWorkerpoolOrderbook[randomIndex]?.order;
+          if (!desiredPriceWorkerpoolOrder) {
+            throw new Error('No Workerpool order found for the desired price');
+          }
+          return desiredPriceWorkerpoolOrder;
+        }),
+    ]);
+
+    const datasetorder = datasetorderForApp || datasetorderForWhitelist;
     if (!datasetorder) {
-      throw new Error('Dataset order not found');
-    }
-
-    // Fetch app order
-    const appOrderbook = await iexec.orderbook.fetchAppOrderbook(
-      WEB3_MAIL_DAPP_ADDRESS,
-      {
-        minTag: ['tee', 'scone'],
-        maxTag: ['tee', 'scone'],
-        workerpool: WORKERPOOL_ADDRESS,
-      }
-    );
-    const appOrder = appOrderbook?.orders[0]?.order;
-    if (!appOrder) {
-      throw new Error('App order not found');
-    }
-
-    const desiredPriceAppOrderbook = appOrderbook.orders.filter(
-      (order) => order.order.appprice <= MAX_DESIRED_APP_ORDER_PRICE
-    );
-    const desiredPriceAppOrder = desiredPriceAppOrderbook[0]?.order;
-    if (!desiredPriceAppOrder) {
-      throw new Error('No App order found for the desired price');
-    }
-
-    // Fetch workerpool order
-    const workerpoolOrderbook = await iexec.orderbook.fetchWorkerpoolOrderbook({
-      workerpool: WORKERPOOL_ADDRESS,
-      app: WEB3_MAIL_DAPP_ADDRESS,
-      dataset: vDatasetAddress,
-      minTag: ['tee', 'scone'],
-      maxTag: ['tee', 'scone'],
-      category: 0,
-    });
-
-    const workerpoolorder = workerpoolOrderbook?.orders[0]?.order;
-    if (!workerpoolorder) {
-      throw new Error('Workerpool order not found');
-    }
-
-    const desiredPriceWorkerpoolOrderbook = workerpoolOrderbook.orders.filter(
-      (order) =>
-        order.order.workerpoolprice <= MAX_DESIRED_WORKERPOOL_ORDER_PRICE
-    );
-    const randomIndex = Math.floor(
-      Math.random() * desiredPriceWorkerpoolOrderbook.length
-    );
-    const desiredPriceWorkerpoolOrder =
-      desiredPriceWorkerpoolOrderbook[randomIndex]?.order;
-    if (!desiredPriceWorkerpoolOrder) {
-      throw new Error('No Workerpool order found for the desired price');
+      throw new Error('No Dataset order found for the desired price');
     }
 
     // Push requester secrets
@@ -146,9 +204,14 @@ export const sendEmail = async ({
       .catch((e) => {
         throw new WorkflowError('Failed to encrypt email content', e);
       });
-    const cid = await ipfs.add(encryptedFile).catch((e) => {
-      throw new WorkflowError('Failed to upload encrypted email content', e);
-    });
+    const cid = await ipfs
+      .add(encryptedFile, {
+        ipfsNode: ipfsNode,
+        ipfsGateway: ipfsGateway,
+      })
+      .catch((e) => {
+        throw new WorkflowError('Failed to upload encrypted email content', e);
+      });
     const multiaddr = `/ipfs/${cid}`;
 
     await iexec.secrets.pushRequesterSecret(
@@ -163,27 +226,29 @@ export const sendEmail = async ({
     );
 
     const requestorderToSign = await iexec.order.createRequestorder({
-      app: WEB3_MAIL_DAPP_ADDRESS,
-      category: desiredPriceWorkerpoolOrder.category,
+      app: vDappAddressOrENS,
+      category: workerpoolorder.category,
       dataset: vDatasetAddress,
-      appmaxprice: desiredPriceAppOrder.appprice,
-      workerpoolmaxprice: desiredPriceWorkerpoolOrder.workerpoolprice,
+      datasetmaxprice: datasetorder.datasetprice,
+      appmaxprice: apporder.appprice,
+      workerpoolmaxprice: workerpoolorder.workerpoolprice,
       tag: ['tee', 'scone'],
-      workerpool: WORKERPOOL_ADDRESS,
+      workerpool: vWorkerpoolAddressOrEns,
       params: {
         iexec_developer_logger: true,
         iexec_secrets: {
           1: requesterSecretId,
         },
+        iexec_args: vLabel,
       },
     });
     const requestorder = await iexec.order.signRequestorder(requestorderToSign);
 
     // Match orders and compute task ID
     const { dealid } = await iexec.order.matchOrders({
-      apporder: desiredPriceAppOrder,
+      apporder: apporder,
       datasetorder: datasetorder,
-      workerpoolorder: desiredPriceWorkerpoolOrder,
+      workerpoolorder: workerpoolorder,
       requestorder: requestorder,
     });
     const taskId = await iexec.deal.computeTaskId(dealid, 0);