+++ tests/unit/fetchMyContacts.test.ts
@@ -6,14 +6,15 @@ import { Wallet } from 'ethers';
 import {
   DATAPROTECTOR_SUBGRAPH_ENDPOINT,
   WEB3_MAIL_DAPP_ADDRESS,
-} from '../../src/config/config';
-import { fetchMyContacts } from '../../dist/web3mail/fetchMyContacts';
+  WHITELIST_SMART_CONTRACT_ADDRESS,
+} from '../../src/config/config.js';
+import { fetchMyContacts } from '../../src/web3mail/fetchMyContacts.js';
 
 describe('fetchMyContacts', () => {
   it('should fetch contacts with the specified page and page size', async () => {
     const graphQLClient = new GraphQLClient(DATAPROTECTOR_SUBGRAPH_ENDPOINT);
     const ethProvider = getWeb3Provider(Wallet.createRandom().privateKey);
-    let iexec = new IExec({
+    const iexec = new IExec({
       ethProvider,
     });
     const MOCK_ORDER = {
@@ -36,34 +37,40 @@ describe('fetchMyContacts', () => {
       status: 'open',
       remaining: 10,
     };
-    const mockFetchDatasetOrderbook: any = jest
-      .fn()
-      .mockImplementation(() => {
-        return Promise.resolve({
-          ok: true,
-          count: 1,
-          nextPage: 1,
-          orders: [MOCK_ORDER],
-        });
+    const mockFetchDatasetOrderbook: any = jest.fn().mockImplementation(() => {
+      return Promise.resolve({
+        ok: true,
+        count: 1,
+        nextPage: 1,
+        orders: [MOCK_ORDER],
       });
+    });
     iexec.orderbook.fetchDatasetOrderbook = mockFetchDatasetOrderbook;
 
     await fetchMyContacts({
       iexec: iexec,
+      dappAddressOrENS: WEB3_MAIL_DAPP_ADDRESS,
+      dappWhitelistAddress: WHITELIST_SMART_CONTRACT_ADDRESS,
       graphQLClient,
-      page: 1,
-      pageSize: 10,
     });
-    const userAddress = await iexec.wallet.getAddress();
+    const userAddress = (await iexec.wallet.getAddress()).toLowerCase();
     expect(iexec.orderbook.fetchDatasetOrderbook).toHaveBeenNthCalledWith(
-      2,
+      1,
       'any',
       {
         app: WEB3_MAIL_DAPP_ADDRESS,
         requester: userAddress,
-        page: 1,
-        pageSize: 10,
+        pageSize: 1000,
       }
     );
-  }, 40_000);
+    expect(iexec.orderbook.fetchDatasetOrderbook).toHaveBeenNthCalledWith(
+      2,
+      'any',
+      {
+        app: WHITELIST_SMART_CONTRACT_ADDRESS.toLowerCase(),
+        requester: userAddress,
+        pageSize: 1000,
+      }
+    );
+  });
 });