+++ CHANGELOG.md
@@ -2,18 +2,36 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.0]
+
+### Added
+
+- added `fetchUserContacts` method allowing to fetch contacts of a given user
+- added `workerpoolAddressOrEns` options to `sendEmail` to specify the workerpool to use (defaults to iExec's production workerpool)
+- added `dataMaxPrice`, `appMaxPrice` and `workerpoolMaxPrice` options to `sendEmail` to specify the maximum price in nRLC to pay to each provider
+- added `label` option to `sendEmail` to tag the web3mail task onchain via iexec_args
+- added `dappAddressOrENS`, `dappWhitelistAddress`, `dataProtectorSubgraph`, `ipfsNode`, `ipfsGateway` configuration options to the constructor
+
+### Changed
+
+- optimized market api requests in `fetchMyContact` and `sendEmail`
+- fixed bug that prevented ethereum address to be used as `dappAddressOrENS`
+- use iExec production workerpool for `sendEmail` if none is specified
+- support for access granted to web3mail dapps whitelist in `sendEmail`
+- migrated from ethers v5 to ethers v6
+
 ## [0.5.2]
 
 ### Added
 
-- add whitelist address as a filter for the restricted app address for fetchMyContacts function
+- support for access granted to web3mail dapps whitelist in `fetchMyContacts`
 - add linter
 
 ## [0.5.1]
 
 ### Changed
 
-- fixed bug that prevented web3mail to work with web3mail dap 0.6.0
+- fixed bug that prevented web3mail to work with web3mail dapp 0.6.0
 
 ## [0.5.0]